refactor(behavior_path_planner): reorganize debug visualization
