+++ planning/behavior_path_planner/src/behavior_path_planner_node.cpp
@@ -14,6 +14,7 @@
 
 #include "behavior_path_planner/behavior_path_planner_node.hpp"
 
+#include "behavior_path_planner/debug_utilities.hpp"
 #include "behavior_path_planner/path_utilities.hpp"
 #include "behavior_path_planner/scene_module/avoidance/avoidance_module.hpp"
 #include "behavior_path_planner/scene_module/lane_change/lane_change_module.hpp"
